requirements.txt
beautifulsoup4==4.12.3
mysql-connector-python==8.3.0
Requests==2.31.0
